add a release process
